EU-115 refactor hostel
